Refactor: 불필요한 렌더링 개선을 위한 컴포넌트화
