+++ src/components/ChartRight/ChartCoinInfo/index.ts
@@ -1,0 +1,2 @@
+import ChartCoinInfo from './ChartCoinInfo';
+export { ChartCoinInfo };
