+++ src/components/ChartRight/ChartContent/ChartContent.styles.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const ChartRefDiv = styled.div`
+  border: 1px solid ${({ theme }) => theme.colors.border1};
+`;
