+++ src/components/ChartRight/ChartContent/index.ts
@@ -1,0 +1,1 @@
+export { ChartContent } from './ChartContent';
